feat: Introduce `StepExecutorPatch` and update specs to enable robust inline asynchronous reactor testing.
