+++ spec/examples/data_pipeline_spec.rb
@@ -31,10 +31,6 @@ RSpec.describe "Data Pipeline - Feature Parity with Elixir" do
   # ============================================================================
 
   describe "Complete ETL Pipeline" do
-    around do |example|
-      Sidekiq::Testing.inline! { example.run }
-    end
-
     let(:sample_users) do
       [
         { "id" => "1", "email" => "alice@example.com", "name" => "Alice Smith", "phone" => "+1-555-0100",
@@ -46,17 +42,16 @@ RSpec.describe "Data Pipeline - Feature Parity with Elixir" do
 
     it "executes complete ETL pipeline with inline execution" do
       # Create reactor instance to access context
-      reactor = UserETLReactor.new
-      result = reactor.run(
-        source_file: "/tmp/users.csv",
-        csv_data: sample_users,
-        output_destinations: %w[database search]
-      )
+      subject = test_reactor(UserETLReactor, inputs: {
+                               source_file: "/tmp/users.csv",
+                               csv_data: sample_users,
+                               output_destinations: %w[database search]
+                             })
 
-      expect(result).to be_a(RubyReactor::Success)
+      expect(subject).to be_success
 
       # The reactor returns :generate_report, so result.value is the report directly
-      report = result.value
+      report = subject.result.value
       expect(report).not_to be_nil
       expect(report[:successful_count]).to eq(3)
       expect(report[:failed_count]).to eq(0)
@@ -64,7 +59,7 @@ RSpec.describe "Data Pipeline - Feature Parity with Elixir" do
       expect(report[:source_count]).to eq(3)
 
       # Access intermediate results from context
-      context = reactor.context
+      context = subject.reactor_instance.context
       transformed = context.intermediate_results[:transform_users]
       expect(transformed).to be_an(Array)
       expect(transformed.length).to eq(3)
