+++ spec/spec_helper.rb
@@ -65,4 +65,8 @@ RSpec.configure do |config|
   config.before do
     allow(RubyReactor::RetryContext).to receive(:calculate_backoff_delay).and_return(0)
   end
+
+  # Apply StepExecutor patch for inline async testing
+  require "ruby_reactor/rspec/step_executor_patch"
+  RubyReactor::Executor::StepExecutor.prepend(RubyReactor::RSpec::StepExecutorPatch)
 end
